renamed _model as _settings
