+++ clients/python/client/osparc/_settings.py
